requirements.txt
fastapi
uvicorn
langchain
langchain-community
sqlalchemy
pydantic
langchain-openai
python-dotenv
langchain-google-genai
pymongo